+++ src/main/java/bob/cloverville/controllers/MemberViewController.java
@@ -4,7 +4,9 @@ import bob.cloverville.Member;
 import bob.cloverville.AppContext;
 import bob.cloverville.MemberService;
 import javafx.fxml.FXML;
+import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.Label;
 import javafx.stage.Stage;
 
@@ -45,14 +47,25 @@ public class MemberViewController {
 
   @FXML
   private void deleteMember() {
-    try{
-      memberService.deleteMember(member);
-      AppContext.get().getDashboardController().loadMembersView();
-    } catch (Exception e){
-      e.printStackTrace();
-    }
-    closeWindow();
+    Alert confirm = new Alert(Alert.AlertType.CONFIRMATION);
+    confirm.setTitle("Confirm Delete");
+    confirm.setHeaderText(null);
+    confirm.setContentText("Are you sure you want to delete this member?");
+
+    confirm.showAndWait().ifPresent(response -> {
+      if (response == ButtonType.OK) {
+        try {
+          memberService.deleteMember(member);
+          AppContext.get().getDashboardController().loadMembersView();
+        } catch (Exception e) {
+          e.printStackTrace();
+        }
+
+        closeWindow();
+      }
+    });
   }
+
 
   @FXML
   private void closeWindow() {
